Backend/requirements.txt
Django==3.0.3
djangorestframework==3.11.0
Markdown==3.2.1
django-filter==2.2.0
psycopg2==2.8.3
django-extensions==2.2.1
shortuuid==0.5.0
uuid==1.30
colorlog==4.1.0